requirements/env38_neuroner.txt
asgiref==3.2.3
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-appconf==1.0.3
django-contrib-comments==1.9.2
django-crispy-forms==1.9.0
django-dotenv==1.4.1
django-imagekit==4.0.2
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
l18n==2018.5
oauthlib==3.1.0
pilkit==2.0
Pillow==6.2.2
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.8
wagtailfontawesome==1.1.4
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
